[Bronze V] Title: 이상한 기호, Time: 52 ms, Memory: 5500 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -8,8 +8,8 @@ namespace testcoding
             using (StreamWriter sw = new StreamWriter(Console.OpenStandardOutput(), bufferSize: 1048576))
             {
                 string[] input = sr.ReadLine().Split();
-                int A = int.Parse(input[0]);
-                int B = int.Parse(input[1]);
+                long A = long.Parse(input[0]);
+                long B = long.Parse(input[1]);
                 sw.WriteLine((A + B) * (A - B));
             }
         }
